refactor(run_context): split storage/query into run_records.py and portability into run_record_portability.py
Step 1: create toolkit/core/run_records.py with:
  - get_run_dir, _run_record_path (path resolution)
  - write_run_record (with Windows-safe rename retry)
  - read_run_record, list_runs, latest_run (query)

Step 2: create toolkit/core/run_record_portability.py with:
  - _load_run_record (portability migration on read)
  - _PORTABLE_RUN_PATH_FIELDS whitelist
  - _migrate_path_value, _migrate_whitelisted_path_fields
  - _root_from_run_dir, _to_pure_path, _is_absolute_path_string

run_context.py becomes thin facade:
  - RunContext class stays here (lifecycle boundary)
  - _now_iso, _duration_seconds, _empty_layer_metrics stay here
  - re-exports of latest_run, list_runs, read_run_record for backward compat
  - no new coupling introduced

No changes to public API, run record JSON format, or test contracts.
